requirements.txt
Django==3.0.7
djangorestframework==3.11.0
django-rest-auth[with_social]==0.9.5
Pygments==2.6.1
pandas==1.0.5
geopy==1.22.0
networkx==2.4